+++ 1859997-1860021-1859905/Cuenta/README.md
@@ -1,0 +1,6 @@
+## main.cpp
+
+* Realizar el switch de las opciones dentro de un while o do - while para permanecer en la interfaz mientras el usuario no desee salir
+* No se estan utilizando los métodos para modiuficar los atributos. Se estan modificando los atributos directamente declarandolos públicos, esta acción no  es recomendable
+
+# Nota: 3.5
