+++ 186014-1860245/Calculadora/README.md
@@ -1,0 +1,13 @@
+## Diagrama de clases
+
+* El diagram de clases se reopresenta mediante un rectangulo con tres secciones onde van el nombre de la clase, los atributos y los métodos y en ese diagram no esta representado de esa forma
+
+## calculadora.cpp
+
+* Cuando realicen la división deben tener en cuenta la que la división entre 0 no se puede hacer
+
+## main.cpp
+
+* Pueden usar funciones para ahorrar lineas de código. En todos las secciones del switch estan escribiendo las mismas lineas
+
+# Nota: 4.5
