+++ 186014-1860245/Cuenta/README.md
@@ -1,0 +1,10 @@
+# Generales
+
+* Organizar mejor el código, cada vez que haya un (;), es recomendable hacer un salto de linea
+
+## menu.cpp
+
+* Cuando no el retiro no sea exitoso se debe mostrar un mensaje
+* Si el usuario no puede ingresar se debe mostrar un mensaje
+
+# Nota: 4.8
