+++ 1866205-1860321-1860255/Cuenta/README.md
@@ -1,0 +1,6 @@
+## Generales
+
+* Organizar mejor el código, cada vez que haya un (;), es recomendable hacer un salto de linea
+* No es necesario la función reingresar desde el menu donde ya ha ingresado
+
+# Nota: 5.0
